backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
<<<<<<< HEAD
=======
click==8.1.3
>>>>>>> happy-branch
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
<<<<<<< HEAD
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.1.8
=======
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
>>>>>>> happy-branch
multidict==6.0.4
openai==0.27.4
parsimonious==0.9.0
protobuf==4.22.3
pycryptodome==3.17
pyrsistent==0.19.3
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
<<<<<<< HEAD
=======
six==1.16.0
>>>>>>> happy-branch
toolz==0.12.0
tqdm==4.65.0
urllib3==1.26.15
web3==6.2.0
websockets==11.0.1
<<<<<<< HEAD
yarl==1.8.2
=======
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
>>>>>>> happy-branch
